amber2sigen/requirements.txt
requests
pycryptodome
python-dateutil
paho-mqtt
